pyproject.toml
[project]
name = "orgst"
version = "0.1.0"
description = "django-ninja project linters template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.12.1",
    "django-ninja>=1.5.3",
    "pillow>=12.1.1",
    "psycopg>=3.3.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.15.1",
]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = [
  "*/migrations/*",
  "*/static/*",
  "*/media/*",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = [
  "E501",  # (opcional) se você preferir não brigar com linhas longas; se quiser, remova
]

[tool.ruff.lint.per-file-ignores]
# Django: imports "expostos" em __init__.py costumam disparar F401
"**/__init__.py" = ["F401"]

# Django Ninja/FastAPI-style: File(...) em default arg (B008) é idiomático
"src/apps/community/views.py" = ["B008"]

# Settings às vezes têm strings longas/constantes
"src/orgst/settings.py" = ["E501"]
"src/orgst/settings/base.py" = ["E501"]
